+++ metrics_utility/exceptions.py
@@ -1,0 +1,8 @@
+class BadShipTarget(Exception):
+    def __init__(self, message):
+        self.name = message
+
+
+class MissingRequiredEnvVar(Exception):
+    def __init__(self, message):
+        self.name = message