+++ CHANGELOG.md
@@ -18,3 +18,5 @@
 
 - gather_automation_controller_billing_data command extension
   Adding ability to read AWS params required for provider billing.
+- adding METRICS_UTILITY_SHIP_TARGET="directory", so the billing data
+  will be stored in a provided directory in daily partitions